# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/companies/{companyId}/job-offers": {
      "get": {
        "operationId": "CompaniesController_getCompanyJobOffers",
        "summary": "Recupera tutte le offerte di una specifica azienda suddivise per stato (conteggi per pendenti/squalificate, dettaglio completo per le attive)",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "description": "UUID dell'azienda",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyJobOffersBreakdownDto"
                }
              }
            }
          }
        },
        "tags": [
          "Companies"
        ]
      }
    }
  },
  "info": {
    "title": "Analizzatore di Ruoli & Job Aggregator API",
    "description": "API Gateway per la ricerca, estrazione e gestione degli annunci di lavoro",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "JobCurriculumDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "jobOfferId": {
            "type": "string"
          },
          "filePath": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "tailoringData": {
            "type": "object",
            "description": "Oggetto JSON di personalizzazione del curriculum (ResumeTailoring)",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "jobOfferId",
          "filePath",
          "explanation",
          "createdAt",
          "updatedAt"
        ]
      },
      "JobStatusHistoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fromStatus": {
            "type": "string",
            "enum": [
              "NEW",
              "SAVED",
              "APPLIED",
              "SCREENING",
              "INTERVIEWING",
              "OFFER",
              "ACCEPTED",
              "REJECTED",
              "ARCHIVED"
            ],
            "nullable": true
          },
          "toStatus": {
            "type": "string",
            "enum": [
              "NEW",
              "SAVED",
              "APPLIED",
              "SCREENING",
              "INTERVIEWING",
              "OFFER",
              "ACCEPTED",
              "REJECTED",
              "ARCHIVED"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "fromStatus",
          "toStatus",
          "createdAt"
        ]
      },
...
      "CompanyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "websiteUrl": {
            "type": "object",
            "nullable": true
          },
          "linkedinUrl": {
            "type": "object",
            "nullable": true
          },
          "industry": {
            "type": "object",
            "nullable": true
          },
          "fundingStage": {
            "type": "object",
            "nullable": true
          },
          "companySizeRange": {
            "type": "object",
            "nullable": true
          },
          "employeeCount": {
            "type": "integer",
            "nullable": true
          },
          "eligibleOffersCount": {
            "type": "integer",
            "nullable": true
          },
          "activeOffersCount": {
            "type": "integer",
            "nullable": true
          },
          "savedOrAppliedCount": {
            "type": "integer",
            "nullable": true
          },
          "newOffersCount": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "JobEvaluationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "jobOfferId": {
            "type": "string"
          },
          "desireMatchScore": {
            "type": "number",
            "description": "Punteggio di aderenza ai desiderata dell utente (0-100)"
          },
          "competenceScore": {
            "type": "number",
            "description": "Punteggio di competenza tecnica (0-100)"
          },
          "overallScore": {
            "type": "number",
            "description": "Punteggio globale bilanciato (0-100)"
          },
          "priority": {
            "type": "string",
            "description": "Livello di priorità: HIGH, MEDIUM, LOW, DISQUALIFIED"
          },
          "status": {
            "type": "string",
            "description": "Stato dell analisi: SUCCESS | UNANALYZABLE"
          },
          "evaluatorModel": {
            "type": "string",
            "description": "Modello AI che ha analizzato l annuncio: GEMINI_3_1_FLASH_LITE | GEMMA_4_12B | UNKNOWN"
          },
          "summary": {
            "type": "object",
            "description": "Sintesi/Riassunto dell annuncio di lavoro generato dall AI",
            "nullable": true
          },
          "desireMatchReasoning": {
            "type": "object",
            "description": "Spiegazione sintetico aderenza desiderata",
            "nullable": true
          },
          "competenceMatch": {
            "type": "string",
            "description": "Valutazione delle competenze dell utente rispetto al ruolo"
          },
          "detailedReasoning": {
            "type": "string",
            "description": "Spiegazione dettagliata dell AI in Markdown"
          },
          "pros": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cons": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "jobOfferId",
          "desireMatchScore",
          "competenceScore",
          "overallScore",
          "priority",
          "status",
          "evaluatorModel",
          "competenceMatch",
          "detailedReasoning",
          "createdAt",
          "updatedAt"
        ]
      },
      "JobOfferDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "externalId": {
            "type": "object",
            "nullable": true
          },
          "source": {
            "type": "string",
            "enum": [
              "ARBEITNOW",
              "REMOTIVE",
              "JOBICY",
              "WE_WORK_REMOTELY"
            ]
          },
          "url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "location": {
            "type": "object",
            "nullable": true
          },
          "remoteType": {
            "type": "string",
            "enum": [
              "REMOTE",
              "HYBRID",
              "ON_SITE",
              "UNSPECIFIED"
            ],
            "default": "UNSPECIFIED"
          },
          "rawDescription": {
            "type": "string"
          },
          "descriptionMarkdown": {
            "type": "object",
            "nullable": true
          },
          "datePosted": {
            "type": "object",
            "nullable": true
          },
          "company": {
            "$ref": "#/components/schemas/CompanyDto"
          },
          "salaryMin": {
            "type": "object",
            "nullable": true
          },
          "salaryMax": {
            "type": "object",
            "nullable": true
          },
          "currency": {
            "type": "object",
            "nullable": true,
            "default": "EUR"
          },
          "contractType": {
            "type": "object",
            "nullable": true
          },
          "roleCategory": {
            "type": "object",
            "nullable": true
          },
          "experienceLevel": {
            "type": "string",
            "enum": [
              "ENTRY_LEVEL",
              "JUNIOR",
              "MID",
              "SENIOR",
              "LEAD",
              "EXECUTIVE",
              "UNSPECIFIED"
            ],
            "default": "UNSPECIFIED"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "SAVED",
              "APPLIED",
              "SCREENING",
              "INTERVIEWING",
              "OFFER",
              "ACCEPTED",
              "REJECTED",
              "ARCHIVED"
            ],
            "default": "NEW"
          },
          "notes": {
            "type": "object",
            "nullable": true
          },
          "evaluation": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/JobEvaluationDto"
              }
            ]
          },
          "curriculum": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/JobCurriculumDto"
              }
            ]
          },
          "freshness": {
            "type": "string",
            "enum": [
              "HOT",
              "RECENT",
              "AGING",
              "OLD"
            ],
            "example": "HOT"
          },
          "statusHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobStatusHistoryDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "url",
          "title",
          "remoteType",
          "rawDescription",
          "company",
          "experienceLevel",
          "status",
          "freshness",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CompanyJobOffersBreakdownDto": {
        "type": "object",
        "properties": {
          "company": {
            "$ref": "#/components/schemas/CompanyDto"
          },
          "counts": {
            "type": "object"
          },
          "offers": {
            "description": "Informazioni complete delle sole offerte idonee (escluse quelle squalificate ed in attesa di valutazione)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobOfferDto"
            }
          }
        },
        "required": [
          "company",
          "counts",
          "offers"
        ]
      }
    }
  }
}